fix(es): parseHistogramLayer key must use float64 for ES scientific notation
Root cause: ES histogram aggregation on long fields returns bucket
keys in scientific notation (e.g. 1.7847756E18 instead of 1784775600000000000).
json.Unmarshal cannot convert '1.7847756E18' into int64 — it silently fails,
returning nil → parseMetricAggResult returns empty series.

Fix: use float64 for the histogram bucket key, then cast to int64.
  json: cannot unmarshal number 1.7847756E18 into int64 ❌
  json: float64 → int64(1.7847756E18) = 1784775600000000000 ✓

Verified against real ES (9.134.106.132:9200):
  Before: SearchLogMetric → 0 series
  After:  SearchLogMetric → 5 series (java-user-service:13850, ...)

+++ extension/observabilitystorageext/provider/elasticsearch/log_reader.go
@@ -217,10 +217,12 @@ func (r *LogReader) parseHistogramLayer(raw map[string]json.RawMessage, labels m
 		}
 	}
 
+	// ES histogram on long fields may return keys in scientific notation
+	// (e.g. 1.7847756E18). Use float64 then convert to int64.
 	var histAgg struct {
 		Buckets []struct {
-			Key      int64 `json:"key"`
-			DocCount int64 `json:"doc_count"`
+			Key      float64 `json:"key"`
+			DocCount int64   `json:"doc_count"`
 		} `json:"buckets"`
 	}
 	if err := json.Unmarshal(histRaw, &histAgg); err != nil {
@@ -233,7 +235,7 @@ func (r *LogReader) parseHistogramLayer(raw map[string]json.RawMessage, labels m
 			continue // skip leading empty buckets
 		}
 		values = append(values, LogMetricValue{
-			TimestampNano: b.Key,
+			TimestampNano: int64(b.Key),
 			Value:         float64(b.DocCount),
 		})
 	}
